put peers function in different file
